refactor: integrate shader in HeightmapTile
